+++ packages/effect-acp/src/protocol.ts
@@ -6,7 +6,7 @@ import * as Ref from "effect/Ref";
 import * as Schema from "effect/Schema";
 import * as Scope from "effect/Scope";
 import * as Stream from "effect/Stream";
-import * as Stdio from "effect/Stdio";
+import { Stdio } from "effect";
 import * as RpcClient from "effect/unstable/rpc/RpcClient";
 import * as RpcClientError from "effect/unstable/rpc/RpcClientError";
 import * as RpcMessage from "effect/unstable/rpc/RpcMessage";